+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 - Complete properties for attributes array keys in closure states.
 - Resolve references for attributes array keys in closure states.
+- Detect unknown properties in state and make methods both inside and outside factory.
 
 ## [1.0.0]
 ### Added